requirements.txt
fastapi 
pydantic 
typing 
openai
uvicorn
langchain
langchain_openai
langchain_core
langchain_chromadb